problems with void MikePositionOrders::updateOpenOrdersByPrice()
added a set container for indexing purposes.
compiles but crashes after pressing a couple of B LMT or other order
buttons and moving price up/down. the lines causing the crash can be
commented out and everything runs smooth.
WHY?!?

To see the crash:
launch program. press 'reprice table'. move the price up or down a bit,
press a couple of order buttons in WidgetTable. move the price to make
the orders fill.